refactor: parser tests
