+++ packages/parser/__tests__/output.test.ts
@@ -1,7 +1,5 @@
 import { describe, expect, test } from 'vitest'
-import { getFixtureProject, parseAndExtract } from './fixture'
-
-const run = parseAndExtract
+import { parseAndExtract } from './fixture'
 
 describe('extract to css output pipeline', () => {
   test('basic usage', () => {
@@ -44,7 +42,7 @@ describe('extract to css output pipeline', () => {
         )
        }
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -189,7 +187,7 @@ describe('extract to css output pipeline', () => {
 
       css({ mx: '3', paddingTop: '4' }, { mx: '10', pt: '6' })
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -247,7 +245,7 @@ describe('extract to css output pipeline', () => {
       )
     }
      `
-    const result = run(code, {
+    const result = parseAndExtract(code, {
       theme: {
         extend: {
           recipes: {
@@ -457,7 +455,7 @@ describe('extract to css output pipeline', () => {
       )
     }
      `
-    const result = run(code, {
+    const result = parseAndExtract(code, {
       theme: {
         extend: {
           recipes: {
@@ -606,7 +604,7 @@ describe('extract to css output pipeline', () => {
     }
 
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -665,7 +663,7 @@ describe('extract to css output pipeline', () => {
         }
     \`
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -792,7 +790,7 @@ describe('extract to css output pipeline', () => {
         color: token(colors.blue.100);
     \`
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -844,7 +842,7 @@ describe('extract to css output pipeline', () => {
        }
      `
 
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -902,7 +900,7 @@ describe('extract to css output pipeline', () => {
         )
        }
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -973,7 +971,7 @@ describe('extract to css output pipeline', () => {
         )
        }
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -1062,7 +1060,7 @@ describe('extract to css output pipeline', () => {
         );
       };
      `
-      const result = run(code, {
+      const result = parseAndExtract(code, {
         theme: {
           extend: {
             semanticTokens: {
@@ -1198,7 +1196,7 @@ describe('extract to css output pipeline', () => {
         )
       }
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -1267,7 +1265,7 @@ describe('extract to css output pipeline', () => {
         )
       }
      `
-    const result = run(code, {
+    const result = parseAndExtract(code, {
       patterns: {
         extend: {
           stack: {
@@ -1343,7 +1341,7 @@ describe('extract to css output pipeline', () => {
       color: var(--colors-purple-100);
     \`
    `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -1440,7 +1438,7 @@ describe('extract to css output pipeline', () => {
         )
       }
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -1563,7 +1561,7 @@ describe('extract to css output pipeline', () => {
         )
        }
      `
-    const { parse, generator } = getFixtureProject(code, {
+    const result = parseAndExtract(code, {
       theme: {
         extend: {
           recipes: {
@@ -1609,8 +1607,8 @@ describe('extract to css output pipeline', () => {
         },
       },
     })
-    const result = parse()!
-    expect(result?.toArray().map(({ box, ...item }) => item)).toMatchInlineSnapshot(`
+
+    expect(result.json).toMatchInlineSnapshot(`
       [
         {
           "data": [
@@ -1696,9 +1694,36 @@ describe('extract to css output pipeline', () => {
         },
       ]
     `)
-    const css = generator.getParserCss(result)!
-    expect(css).toMatchInlineSnapshot(`
-      "@layer utilities {
+
+    expect(result.css).toMatchInlineSnapshot(`
+      "@layer recipes {
+        .button--size_md {
+          padding: var(--spacing-4);
+          border-radius: var(--radii-md)
+          }
+
+        .button--variant_danger {
+          color: var(--colors-white);
+          background-color: var(--colors-red-500)
+          }
+
+        .anotherButton--spacing_sm {
+          padding: var(--spacing-2);
+          border-radius: var(--radii-sm)
+          }
+
+        .complexButton--color_blue {
+          color: var(--colors-blue-500)
+          }
+
+        @layer _base {
+          .button {
+            font-size: var(--font-sizes-lg)
+              }
+          }
+      }
+
+      @layer utilities {
         .mt_40px {
           margin-top: 40px
           }
@@ -1725,33 +1750,6 @@ describe('extract to css output pipeline', () => {
 
         .gap_10px {
           gap: 10px
-          }
-      }
-
-      @layer recipes {
-        .button--size_md {
-          padding: var(--spacing-4);
-          border-radius: var(--radii-md)
-          }
-
-        .button--variant_danger {
-          color: var(--colors-white);
-          background-color: var(--colors-red-500)
-          }
-
-        .anotherButton--spacing_sm {
-          padding: var(--spacing-2);
-          border-radius: var(--radii-sm)
-          }
-
-        .complexButton--color_blue {
-          color: var(--colors-blue-500)
-          }
-
-        @layer _base {
-          .button {
-            font-size: var(--font-sizes-lg)
-              }
           }
       }"
     `)
@@ -1784,7 +1782,7 @@ describe('extract to css output pipeline', () => {
     };
 
      `
-    const { parse, generator } = getFixtureProject(code, {
+    const result = parseAndExtract(code, {
       outdir: 'styled-system',
       jsxFactory: 'styled',
       theme: {
@@ -1798,8 +1796,8 @@ describe('extract to css output pipeline', () => {
         },
       },
     })
-    const result = parse()!
-    expect(result?.toArray().map(({ box, ...item }) => item)).toMatchInlineSnapshot(`
+
+    expect(result.json).toMatchInlineSnapshot(`
       [
         {
           "data": [
@@ -1847,8 +1845,8 @@ describe('extract to css output pipeline', () => {
         },
       ]
     `)
-    const css = generator.getParserCss(result)!
-    expect(css).toMatchInlineSnapshot(`
+
+    expect(result.css).toMatchInlineSnapshot(`
       "@layer utilities {
         .bg_tomato {
           background-color: tomato
@@ -1886,7 +1884,7 @@ describe('extract to css output pipeline', () => {
       "& h2": paragraphSpacingStyle,
     });`
 
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -1983,7 +1981,7 @@ describe('extract to css output pipeline', () => {
     })
      `
 
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2069,7 +2067,7 @@ describe('extract to css output pipeline', () => {
       );
     }
      `
-    const result = run(code, {
+    const result = parseAndExtract(code, {
       outdir: 'anywhere',
       importMap: {
         css: 'controlled-import-map/css',
@@ -2078,6 +2076,7 @@ describe('extract to css output pipeline', () => {
         jsx: 'controlled-import-map',
       },
     })
+
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2120,29 +2119,7 @@ describe('extract to css output pipeline', () => {
     `)
 
     expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .mx_3 {
-          margin-inline: var(--spacing-3)
-          }
-
-        .d_flex {
-          display: flex
-          }
-
-        .flex_column {
-          flex-direction: column
-          }
-
-        .gap_10px {
-          gap: 10px
-          }
-
-        .text_red {
-          color: red
-          }
-      }
-
-      @layer recipes {
+      "@layer recipes {
         .buttonStyle--size_md {
           height: 3rem;
           min-width: 3rem;
@@ -2170,6 +2147,28 @@ describe('extract to css output pipeline', () => {
             justify-content: center
               }
           }
+      }
+
+      @layer utilities {
+        .mx_3 {
+          margin-inline: var(--spacing-3)
+          }
+
+        .d_flex {
+          display: flex
+          }
+
+        .flex_column {
+          flex-direction: column
+          }
+
+        .gap_10px {
+          gap: 10px
+          }
+
+        .text_red {
+          color: red
+          }
       }"
     `)
   })
@@ -2185,7 +2184,7 @@ describe('extract to css output pipeline', () => {
          }
        `
 
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2222,7 +2221,7 @@ describe('extract to css output pipeline', () => {
          }
        `
 
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2269,7 +2268,7 @@ describe('extract to css output pipeline', () => {
          }
        `
 
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2355,7 +2354,7 @@ describe('extract to css output pipeline', () => {
     }
 
      `
-    const result = run(code, {
+    const result = parseAndExtract(code, {
       theme: {
         extend: {
           recipes: {
@@ -2420,25 +2419,7 @@ describe('extract to css output pipeline', () => {
     `)
 
     expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .text_amber\\\\.400 {
-          color: var(--colors-amber-400)
-          }
-
-        [data-theme=dark] .dark\\\\:text_sky\\\\.300, .dark .dark\\\\:text_sky\\\\.300, .dark\\\\:text_sky\\\\.300.dark, .dark\\\\:text_sky\\\\.300[data-theme=dark] {
-          color: var(--colors-sky-300)
-              }
-
-        .hover\\\\:text_amber\\\\.500:is(:hover, [data-hover]) {
-          color: var(--colors-amber-500)
-              }
-
-        [data-theme=dark] .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover]), .dark .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover]), .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover]).dark, .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover])[data-theme=dark] {
-          color: var(--colors-sky-200)
-                  }
-      }
-
-      @layer recipes {
+      "@layer recipes {
         .button--size_md {
           border-radius: var(--radii-md)
           }
@@ -2457,6 +2438,24 @@ describe('extract to css output pipeline', () => {
             background: var(--colors-red-900)
               }
           }
+      }
+
+      @layer utilities {
+        .text_amber\\\\.400 {
+          color: var(--colors-amber-400)
+          }
+
+        [data-theme=dark] .dark\\\\:text_sky\\\\.300, .dark .dark\\\\:text_sky\\\\.300, .dark\\\\:text_sky\\\\.300.dark, .dark\\\\:text_sky\\\\.300[data-theme=dark] {
+          color: var(--colors-sky-300)
+              }
+
+        .hover\\\\:text_amber\\\\.500:is(:hover, [data-hover]) {
+          color: var(--colors-amber-500)
+              }
+
+        [data-theme=dark] .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover]), .dark .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover]), .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover]).dark, .hover\\\\:dark\\\\:text_sky\\\\.200:is(:hover, [data-hover])[data-theme=dark] {
+          color: var(--colors-sky-200)
+                  }
       }"
     `)
   })
@@ -2475,7 +2474,7 @@ describe('extract to css output pipeline', () => {
     }
 
      `
-    const result = run(code, {
+    const result = parseAndExtract(code, {
       theme: {
         extend: {
           recipes: {
@@ -2524,19 +2523,7 @@ describe('extract to css output pipeline', () => {
     `)
 
     expect(result.css).toMatchInlineSnapshot(`
-      "@layer utilities {
-        .fs_2 {
-          font-size: 2px
-          }
-
-        @media screen and (min-width: 40em) {
-          .sm\\\\:fs_5 {
-            font-size: 5px
-          }
-              }
-      }
-
-      @layer recipes {
+      "@layer recipes {
         .card--size_sm {
           border-radius: var(--radii-sm);
           padding: var(--spacing-2);
@@ -2554,6 +2541,18 @@ describe('extract to css output pipeline', () => {
                       }
                   }
           }
+      }
+
+      @layer utilities {
+        .fs_2 {
+          font-size: 2px
+          }
+
+        @media screen and (min-width: 40em) {
+          .sm\\\\:fs_5 {
+            font-size: 5px
+          }
+              }
       }"
     `)
   })
@@ -2571,7 +2570,7 @@ describe('extract to css output pipeline', () => {
       );
     };
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2634,7 +2633,7 @@ describe('extract to css output pipeline', () => {
       }
     })
      `
-    const result = run(code)
+    const result = parseAndExtract(code)
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2686,7 +2685,7 @@ describe('extract to css output pipeline', () => {
       return <CopyButton content="https://www.buymeacoffee.com/grizzlycodes" />
     }
      `
-    const result = run(code, { strictTokens: true })
+    const result = parseAndExtract(code, { strictTokens: true })
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
@@ -2714,7 +2713,7 @@ describe('extract to css output pipeline', () => {
       bgColor: '[rgb(51 155 240)]',
     })
      `
-    const result = run(code, { strictTokens: true })
+    const result = parseAndExtract(code, { strictTokens: true })
     expect(result.json).toMatchInlineSnapshot(`
       [
         {
